backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pyodbc==5.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
reportlab==4.2.5
bcrypt==4.2.1
